refactor: remove all monaco code to common folder
