+++ src/common/index.ts
@@ -2,6 +2,5 @@ export * from './customError';
 export * from './debounceThrottle';
 export * from './debug';
 export * from './pureObject';
-export * from './editor';
 export * from './base64';
 export * from './crypto';
