+++ src/common/monaco/tokenlizer.ts
@@ -1,0 +1,47 @@
+import { executeActions, SearchAction } from './';
+
+export const buildSearchToken = (lines: Array<{ lineNumber: number; lineContent: string }>) => {
+  const commands = lines.filter(({ lineContent }) => executeActions.regexp.test(lineContent));
+
+  return commands.map(({ lineContent, lineNumber }, index, commands) => {
+    const rawCmd = lineContent.split(/[/\s]+/);
+    const method = rawCmd[0]?.toUpperCase();
+    const indexName = rawCmd[1]?.startsWith('_') ? undefined : rawCmd[1];
+    const path = rawCmd.slice(indexName ? 2 : 1, rawCmd.length).join('/');
+    const nexCommandLineNumber = commands[index + 1]?.lineNumber
+      ? commands[index + 1]?.lineNumber - 1
+      : lines.length;
+
+    const endLineNumber =
+      lines
+        .slice(lineNumber, nexCommandLineNumber)
+        .reverse()
+        .find(({ lineContent }) => lineContent.trim().endsWith('}'))?.lineNumber || lineNumber;
+
+    const qdsl = lines
+      .slice(lineNumber, endLineNumber)
+      .map(({ lineContent }) => lineContent)
+      .join('\n');
+
+    return {
+      qdsl,
+      method,
+      index: indexName,
+      path,
+      actionPosition: {
+        startLineNumber: lineNumber,
+        endLineNumber: lineNumber,
+        startColumn: 1,
+        endColumn: lineContent.length,
+      },
+      qdslPosition: qdsl
+        ? {
+            startLineNumber: lineNumber + 1,
+            startColumn: 1,
+            endLineNumber,
+            endColumn: lines[endLineNumber].lineContent.length,
+          }
+        : null,
+    } as SearchAction;
+  });
+};
